requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pdfplumber>=0.10.0
requests>=2.31.0
python-fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
rapidfuzz>=3.0.0
openpyxl>=3.1.0
